fix(user): refactor FilamentPanel usage and add FilamentServiceProvider
- update dependency filament/filament to ^4.2
- update dependency driftingly/rector-laravel to ^2.1.3
