+++ bootstrap/providers.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 use App\Providers\AppServiceProvider;
 use App\Providers\EventServiceProvider;
 use App\Providers\Filament\AdminPanelProvider;
+use App\Providers\Filament\FilamentServiceProvider;
 use App\Providers\Filament\GuestPanelProvider;
 use App\Providers\Filament\PartnerPanelProvider;
 use App\Providers\Filament\UserPanelProvider;
@@ -12,6 +13,7 @@ use App\Providers\RouteServiceProvider;
 
 return [
     AppServiceProvider::class,
+    FilamentServiceProvider::class,
     EventServiceProvider::class,
     AdminPanelProvider::class,
     GuestPanelProvider::class,
