+++ composer.json
@@ -9,7 +9,7 @@
     "license": "MIT",
     "require": {
         "php": "^8.3",
-        "filament/filament": "^4.0",
+        "filament/filament": "^4.2",
         "guzzlehttp/guzzle": "^7.10.0",
         "he4rt/authentication": ">=1",
         "he4rt/badge": "^1.0.0",
@@ -33,7 +33,7 @@
     },
     "require-dev": {
         "barryvdh/laravel-ide-helper": "^3.6.0",
-        "driftingly/rector-laravel": "^2.1.2",
+        "driftingly/rector-laravel": "^2.1.3",
         "fakerphp/faker": "^1.24.1",
         "larastan/larastan": "^3.8.0",
         "laravel/pail": "^1.2.3",
